support multi dialect tests, break up tests more rationally
