+++ packages/malloy-db-test/src/join.spec.ts
@@ -12,10 +12,8 @@
  */
 /* eslint-disable no-console */
 
-import { BigQueryConnection } from "@malloy-lang/db-bigquery";
 import * as malloy from "@malloy-lang/malloy";
-import * as util from "util";
-import * as fs from "fs";
+import { getRuntimes } from "./runtimes";
 
 const joinModelText = `
 export define aircraft_models is ('malloytest.aircraft_models'
@@ -47,44 +45,33 @@ export define aircraft is ('malloytest.aircraft'
 )
 `;
 
+// const runtimes = getRuntimes(["bigquery"]);
+const runtimes = getRuntimes();
+
+const models = new Map<string, malloy.ModelRuntimeRequest>();
+runtimes.forEach((runtime, key) => {
+  models.set(key, runtime.makeModel(joinModelText));
+});
+
 describe("join expression tests", () => {
-  let bq: BigQueryConnection;
-  let runtime: malloy.Runtime;
-  let model: malloy.ModelRuntimeRequest;
-
-  beforeAll(() => {
-    bq = new BigQueryConnection("test");
-    const files = {
-      readUrl: async (url: malloy.Url) => {
-        const filePath = url.toString().replace(/^file:\/\//, "");
-        return await util.promisify(fs.readFile)(filePath, "utf8");
-      },
-    };
-    runtime = new malloy.Runtime({
-      urls: files,
-      schemas: bq,
-      connections: bq,
-    });
-    model = runtime.makeModel(joinModelText);
-  });
-
-  it("model post join", async () => {
-    const result = await model
-      .makeQuery(
-        `
+  models.forEach((model, database) => {
+    it(`model post join - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore aircraft joins aircraft_models on aircraft_model_code | reduce
         aircraft_count,
         aircraft_models.model_count
       `
-      )
-      .run();
-    expect(result.getData().toObject()[0].model_count).toBe(1416);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].model_count).toBe(1416);
+    });
 
-  it("model: join fact table query", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: join fact table query - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore aircraft_models
         joins am_facts is (
           aircraft_models  | reduce
@@ -96,15 +83,15 @@ describe("join expression tests", () => {
         order by 2 desc
         limit 1
     `
-      )
-      .run();
-    expect(result.getData().toObject()[0].num_models).toBe(1147);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].num_models).toBe(1147);
+    });
 
-  it("model: explore based on query", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: explore based on query - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore (
             aircraft_models  | reduce
             m is manufacturer,
@@ -115,15 +102,15 @@ describe("join expression tests", () => {
         order by 2 desc
         limit 1
         `
-      )
-      .run();
-    expect(result.getData().toObject()[0].num_models).toBe(1147);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].num_models).toBe(1147);
+    });
 
-  it("model: funnel - merge two queries", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: funnel - merge two queries - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore (
           aircraft_models  | reduce
             m is manufacturer,
@@ -141,16 +128,16 @@ describe("join expression tests", () => {
         order by 2 desc
         limit 1
         `
-      )
-      .run();
-    expect(result.getData().toObject()[0].num_models).toBe(1147);
-    expect(result.getData().toObject()[0].total_seats).toBe(252771);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].num_models).toBe(1147);
+      expect(result.getData().toObject()[0].total_seats).toBe(252771);
+    });
 
-  it("model: modeled funnel", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: modeled funnel - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore (aircraft_models | manufacturer_models)
         joins seats is (aircraft_models | manufacturer_seats)
           on manufacturer
@@ -161,16 +148,16 @@ describe("join expression tests", () => {
         order by 2 desc
         limit 1
         `
-      )
-      .run();
-    expect(result.getData().toObject()[0].num_models).toBe(1147);
-    expect(result.getData().toObject()[0].total_seats).toBe(252771);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].num_models).toBe(1147);
+      expect(result.getData().toObject()[0].total_seats).toBe(252771);
+    });
 
-  it("model: modeled funnel", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: modeled funnel - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore funnel
       | project
         manufacturer,
@@ -179,33 +166,34 @@ describe("join expression tests", () => {
         order by 2 desc
         limit 1
         `
-      )
-      .run();
-    expect(result.getData().toObject()[0].num_models).toBe(1147);
-    expect(result.getData().toObject()[0].total_seats).toBe(252771);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].num_models).toBe(1147);
+      expect(result.getData().toObject()[0].total_seats).toBe(252771);
+    });
 
-  it("model: double_pipe", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: double_pipe - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore
        (aircraft_models | reduce manufacturer, f is count(*) | reduce f_sum is f.sum())
       | project f_sum2 is f_sum+1
     `
-      )
-      .run();
-    expect(result.getData().toObject()[0].f_sum2).toBe(60462);
-  });
+        )
+        .run();
+      expect(result.getData().toObject()[0].f_sum2).toBe(60462);
+    });
 
-  it("model: double_pipe2", async () => {
-    const result = await model
-      .makeQuery(
-        `
+    it(`model: double_pipe2 - ${database}`, async () => {
+      const result = await model
+        .makeQuery(
+          `
       explore pipe | project f_sum2 is f_sum+1
       `
-      )
-      .run();
-    expect(result.getData().toObject()[0].f_sum2).toBe(60462);
+        )
+        .run();
+      expect(result.getData().toObject()[0].f_sum2).toBe(60462);
+    });
   });
 });
